KidneyDisease-Classification/requirements.txt
tensorflow>=2.12
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
Pillow>=9.5
PyYAML>=6.0
mlflow>=2.10
python-dotenv>=1.0
fastapi>=0.110
uvicorn>=0.23
python-multipart>=0.0.9
Jinja2>=3.1
